Moved creation of tmp folders to main
